pyproject.toml
[project]
name = "ai-database-query-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Ahmad10Raza",email = "rjaahmad60@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "mysql-connector-python (>=9.4.0,<10.0.0)",
    "requests (>=2.32.4,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
